requirements.txt
asgiref==3.4.1
cairocffi==1.3.0
CairoSVG==2.5.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cssselect2==0.4.1
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.2
django-admin-interface==0.17.3
django-colorfield==0.4.5
django-config==0.1.4
django-countries==7.2.1
django-crispy-forms==1.13.0
django-filter==21.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-heroku==0.0.0
django-import-export==2.6.1
django-storages==1.12.2
django-weasyprint==1.1.0.post1
et-xmlfile==1.1.0
gunicorn==20.1.0
html5lib==1.1
idna==3.3
install==1.3.4
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
Pillow==8.4.0
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.3.0
pyphen==0.11.0
python-decouple==3.5
pytz==2021.3
PyYAML==6.0
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
tablib==3.0.0
tinycss2==1.1.0
twilio==7.1.0
urllib3==1.26.7
WeasyPrint==52.5
webencodings==0.5.1
whitenoise==5.3.0
xlrd==2.0.1
xlwt==1.3.0
